Fix access control handling

+++ hawkbit-repository/hawkbit-repository-api/src/main/java/org/eclipse/hawkbit/repository/DeploymentManagement.java
@@ -413,5 +413,5 @@ public interface DeploymentManagement extends PermissionSupport {
     void cancelActionsForDistributionSet(final ActionCancellationType cancelationType, final DistributionSet set);
 
     @PreAuthorize(SpringEvalExpressions.IS_SYSTEM_CODE)
-    void handleMaxAssignmentsExceeded(Long targetId, Integer requested, AssignmentQuotaExceededException ex);
+    void handleMaxAssignmentsExceeded(Long targetId, Long requested, AssignmentQuotaExceededException ex);
 }
