+++ hawkbit-repository/hawkbit-repository-jpa/src/main/java/org/eclipse/hawkbit/repository/jpa/JpaRolloutExecutor.java
@@ -858,7 +858,7 @@ public class JpaRolloutExecutor implements RolloutExecutor {
      * @param target the target
      * @param requested number of actions to check
      */
-    private void assertActionsPerTargetQuota(final Target target, final int requested) {
+    private void assertActionsPerTargetQuota(final Target target, final long requested) {
         final int quota = quotaManagement.getMaxActionsPerTarget();
         try {
             QuotaHelper.assertAssignmentQuota(target.getId(), requested, quota, Action.class, Target.class, actionRepository::countByTargetId);
